Modernize wayland_protocol_codegen
wayland_protocol_codegen.go was originally created as modified fork of
the Soong "gensrcs" module from build/soong/genrule/genrule.go, back in
2017. genrule.go has since changed substantially, and this fork was
never updated.

This patch updates our fork to more closely match the current version of
genrule.go, which includes the ability to convert to use the Bazel build
system.

wayland_protocol_codegen.go remains a fork as the gensrcs module does
not support the one feature we needed (better control over output
filename generation), and would be tricky to modify.

As part of the update, we gained the ability to write a unit test of the
Soong build module, so one has been added. While it doesn't cover all of
the underlying functionality provided by the forked code, it does test
the key bits, and should be enough.

With the upcoming switch to Bazel, the need for a messy fork will go
away. Bazel makes it easier to define build custom rules, and this patch
defines one to be used after the conversion, along with a Bazel test
suite.

Bug: 144185560
Test: m libwayland_extension_client_protocols
Test: b test //external/wayland-protocols/bazel:gensrcs_tests
Test: b build //external/wayland-protocols:libwayland_extension_client_protocols --config=android

Change-Id: I43bd17a897d487f55493425ca9215bb1d8e7fc70