Add overlap for frame_type enum
As part of the shelf/caption insets separation project, we need to
place caption(header) in front of(z-axis) the window area instead of
putting it above(y-axis) the area.

In order to archieve that we are adding a new type overlap for
frame_type, which means that the frame is overlapping with the surface.

Bug: b/282627319
Test: Manual testing to make sure it works when frame type is set to OVERLAY
Change-Id: Ie9107d7e1ec5f30a2f2413146b5676a5292d800c
(cherry picked from commit ef0aaf5a23b578f308384f5f93638bab45acc8b6)