Get HostToolInfo from direct dependency instead of PrebuiltGetPreferred
PrebuiltGetPreferred uses WalkDepsProxy, which pessimizes incremental
analysis.  It is only used to manually handle HostToolProviderInfo  of
prebuilts that replace sources in cases where the dependency was added
by a FinalDepsMutator, too late for the Prebuilts mutators to replace
the dependency.

Get the HostToolInfo from the direct dep, which has already propagated
the prebuilt's HostToolInfo if necessary.

Bug: 457275077
Test: no change to generated ninja files
Change-Id: Ie7ffceb644c2fa4bd6bdcb9b2cd0acd945f3f50d